# Faster hex encoding