--- a/Cargo.toml
+++ b/Cargo.toml
@@ -32,11 +32,23 @@ typenum = { version = "1.16", features = ["const-generics"] }
 const-default = { version = "1", optional = true, default-features = false }
 serde = { version = "1.0", optional = true, default-features = false }
 zeroize = { version = "1", optional = true, default-features = false }
+faster-hex = { version = "0.8", optional = true, default-features = false }
 
 [dev-dependencies]
 # this can't yet be made optional, see https://github.com/rust-lang/cargo/issues/1596
 serde_json = "1.0"
 bincode = "1.0"
+criterion = { version = "0.5", features = ["html_reports"] }
+rand = "0.8"
+
+[[bench]]
+name = "hex"
+harness = false
+
+[profile.bench]
+opt-level = 3
+lto = 'fat'
+codegen-units = 1
 
 [package.metadata.docs.rs]
 # all but "internals", don't show those on docs.rs