Optimizes the standard encoding, plus adds support for using the `faster-hex` crate for SIMD-accelerated encoding.

<details>
<summary>Expand for benchmarks table</summary>

| N | Before |  After | Change |
|---:|---|---|:---:|
| 1 | 16.220 ns | 16.010 ns | -1.30% |
| 2 | 17.035 ns | 16.019 ns | -5.96% |
| 4 | 17.765 ns | 17.000 ns | -4.31% |
| 8 | 18.798 ns | 18.640 ns | -0.84% |
| 12 | 21.002 ns | 20.106 ns | -4.27% |
| 15 | 22.648 ns | 21.692 ns | -4.22% |
| 16 | 22.495 ns | 15.583 ns | -30.73% |
| 32 | 32.556 ns | 16.930 ns | -48.00% |
| 64 | 50.232 ns | 19.022 ns | -62.13% |
| 100 | 70.175 ns | 23.459 ns | -66.57% |
| 128 | 87.228 ns | 20.423 ns | -76.59% |
| 160 | 105.057 ns | 22.228 ns | -78.84% |
| 255 | 156.836 ns | 34.361 ns | -78.09% |
| 256 | 182.728 ns | 25.557 ns | -86.01% |
| 500 | 293.572 ns | 41.154 ns | -85.98% |
| 512 | 352.274 ns | 38.409 ns | -89.10% |
| 900 | 610.794 ns | 63.040 ns | -89.68% |
| 1023 | 585.836 ns | 74.845 ns | -87.22% |
| 1024 | 586.594 ns | 65.048 ns | -88.91% |
| 1025 | 505.389 ns | 72.909 ns | -85.57% |
| 2048 | 975.273 ns | 110.846 ns | -88.63% |
| 4096 | 1945.770 ns | 201.676 ns | -89.64% |
| 5000 | 2343.561 ns | 250.155 ns | -89.33% |
| 10000 | 4698.586 ns | 509.036 ns | -89.17% |

</details>

It's worth noting at least 10-16ns of those benchmarks seems to be spent on non-encoding things. In fact it was made worse when I upgraded to Rust 1.72.0 today, but only in terms of that base offset.

I'm not super happy about the `unsafe` usage here, but when mixing the fmt precision in like that LLVM has a harder time reasoning with it.