@novacrazy 2023-08-25
Re-ran benchmarks with Rust 1.73.0-beta, which includes an LLVM upgrade:

<details>
<summary>Expand for benchmarks table</summary>

| N | Before |  After | Change |
|---:|---|---|:---:|
| 1 | 9.153 ns | 9.202 ns | 0.53% |
| 2 | 9.820 ns | 9.293 ns | -5.37% |
| 4 | 11.704 ns | 9.952 ns | -14.97% |
| 8 | 13.164 ns | 11.583 ns | -12.01% |
| 12 | 13.737 ns | 13.645 ns | -0.67% |
| 15 | 15.083 ns | 14.719 ns | -2.42% |
| 16 | 16.915 ns | 12.643 ns | -25.26% |
| 32 | 28.896 ns | 14.379 ns | -50.24% |
| 64 | 47.562 ns | 15.429 ns | -67.56% |
| 100 | 75.660 ns | 19.488 ns | -74.24% |
| 128 | 83.917 ns | 18.227 ns | -78.28% |
| 160 | 109.399 ns | 23.072 ns | -78.91% |
| 255 | 156.394 ns | 32.404 ns | -79.28% |
| 256 | 159.777 ns | 24.415 ns | -84.72% |
| 500 | 288.924 ns | 39.826 ns | -86.22% |
| 512 | 296.679 ns | 36.177 ns | -87.81% |
| 900 | 548.176 ns | 58.511 ns | -89.33% |
| 1023 | 583.620 ns | 73.875 ns | -87.34% |
| 1024 | 614.904 ns | 63.397 ns | -89.69% |
| 1025 | 566.742 ns | 68.560 ns | -87.90% |
| 2048 | 1084.768 ns | 113.175 ns | -89.57% |
| 4096 | 2146.710 ns | 198.601 ns | -90.75% |
| 5000 | 3290.758 ns | 245.378 ns | -92.54% |
| 10000 | 5580.791 ns | 488.860 ns | -91.24% |

</details>

Old version got slower, new version gets faster. Neat.